pyproject.toml
[project]
name = "solana-mcp"
version = "0.1.0"
description = "Solana MCP Server - A FastMCP server for interacting with the Solana blockchain"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.6.0",
    "requests>=2.32.3",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.28.0",
    "fastmcp>=2.2.5",
    "pydantic>=2.0.0",
]
